+++ src/theory/strings/strings_entail.h
@@ -33,6 +33,225 @@ namespace strings {
 class StringsEntail
 {
  public:
+
+/** can constant contain list
+ * return true if constant c can contain the list l in order
+ * firstc/lastc store which indices in l were used to determine the return
+ * value.
+ *   (This is typically used when this function returns false, for minimizing
+ * explanations)
+ *
+ * For example:
+ *   canConstantContainList( "abc", { x, "c", y } ) returns true
+ *     firstc/lastc are updated to 1/1
+ *   canConstantContainList( "abc", { x, "d", y } ) returns false
+ *     firstc/lastc are updated to 1/1
+ *   canConstantContainList( "abcdef", { x, "b", y, "a", z, "c", w }
+ *     returns false
+ *     firstc/lastc are updated to 1/3
+ *   canConstantContainList( "abcdef", { x, "b", y, "e", z, "c", w }
+ *     returns false
+ *     firstc/lastc are updated to 1/5
+ */
+static bool canConstantContainList(Node c,
+                            std::vector<Node>& l,
+                            int& firstc,
+                            int& lastc);
+/** can constant contain concat
+ * same as above but with n = str.++( l ) instead of l
+ */
+static bool canConstantContainConcat(Node c, Node n, int& firstc, int& lastc);
+
+/** strip symbolic length
+ *
+ * This function strips off components of n1 whose length is less than
+ * or equal to argument curr, and stores them in nr. The direction
+ * dir determines whether the components are removed from the start
+ * or end of n1.
+ *
+ * In detail, this function updates n1 to n1' such that:
+ *   If dir=1,
+ *     n1 = str.++( nr, n1' )
+ *   If dir=-1
+ *     n1 = str.++( n1', nr )
+ * It updates curr to curr' such that:
+ *   curr' = curr - str.len( str.++( nr ) ), and
+ *   curr' >= 0
+ * where the latter fact is determined by checkArithEntail.
+ *
+ * This function returns true if n1 is modified.
+ *
+ * For example:
+ *
+ *  stripSymbolicLength( { x, "abc", y }, {}, 1, str.len(x)+1 )
+ *    returns true
+ *    n1 is updated to { "bc", y }
+ *    nr is updated to { x, "a" }
+ *    curr is updated to 0   *
+ *
+ * stripSymbolicLength( { x, "abc", y }, {}, 1, str.len(x)-1 )
+ *    returns false
+ *
+ *  stripSymbolicLength( { y, "abc", x }, {}, 1, str.len(x)+1 )
+ *    returns false
+ *
+ *  stripSymbolicLength( { x, "abc", y }, {}, -1, 2*str.len(y)+4 )
+ *    returns true
+ *    n1 is updated to { x }
+ *    nr is updated to { "abc", y }
+ *    curr is updated to str.len(y)+1
+ */
+static bool stripSymbolicLength(std::vector<Node>& n1,
+                         std::vector<Node>& nr,
+                         int dir,
+                         Node& curr);
+/** component contains
+ * This function is used when rewriting str.contains( t1, t2 ), where
+ * n1 is the vector form of t1
+ * n2 is the vector form of t2
+ *
+ * If this function returns n>=0 for some n, then
+ *    n1 = { x1...x{n-1} xn...x{n+s} x{n+s+1}...xm },
+ *    n2 = { y1...ys },
+ *    y1 is a suffix of xn,
+ *    y2...y{s-1} = x{n+1}...x{n+s-1}, and
+ *    ys is a prefix of x{n+s}
+ * Otherwise it returns -1.
+ *
+ * This function may update n1 if computeRemainder = true.
+ *   We maintain the invariant that the resulting value n1'
+ *   of n1 after this function is such that:
+ *     n1 = str.++( nb, n1', ne )
+ * The vectors nb and ne have the following properties.
+ * If computeRemainder = true, then
+ *   If remainderDir != -1, then
+ *     ne is { x{n+s}' x{n+s+1}...xm }
+ *     where x{n+s} = str.++( ys, x{n+s}' ).
+ *   If remainderDir != 1, then
+ *     nb is { x1, ..., x{n-1}, xn' }
+ *     where xn = str.++( xn', y1 ).
+ *
+ * For example:
+ *
+ * componentContains({ x, "abc", x }, { "b" }, {}, true, 0)
+ *   returns 1,
+ *   n1 is updated to { "b" },
+ *   nb is updated to { x, "a" },
+ *   ne is updated to { "c", x }
+ *
+ * componentContains({ x, "abc", x }, { "b" }, {}, true, 1)
+ *   returns 1,
+ *   n1 is updated to { x, "ab" },
+ *   ne is updated to { "c", x }
+ *
+ * componentContains({ y, z, "abc", x, "def" }, { "c", x, "de" }, {}, true, 1)
+ *   returns 2,
+ *   n1 is updated to { y, z, "abc", x, "de" },
+ *   ne is updated to { "f" }
+ *
+ * componentContains({ y, "abc", x, "def" }, { "c", x, "de" }, {}, true, -1)
+ *   returns 1,
+ *   n1 is updated to { "c", x, "def" },
+ *   nb is updated to { y, "ab" }
+ */
+static int componentContains(std::vector<Node>& n1,
+                      std::vector<Node>& n2,
+                      std::vector<Node>& nb,
+                      std::vector<Node>& ne,
+                      bool computeRemainder = false,
+                      int remainderDir = 0);
+/** component contains base
+ *
+ * This function is a helper for the above function.
+ *
+ * It returns true if n2 is contained in n1 with the following
+ * restrictions:
+ *   If dir=1, then n2 must be a suffix of n1.
+ *   If dir=-1, then n2 must be a prefix of n1.
+ *
+ * If computeRemainder is true, then n1rb and n1re are
+ * updated such that :
+ *   n1 = str.++( n1rb, n2, n1re )
+ * where a null value of n1rb and n1re indicates the
+ * empty string.
+ *
+ * For example:
+ *
+ * componentContainsBase("cabe", "ab", n1rb, n1re, 1, false)
+ *   returns false.
+ *
+ * componentContainsBase("cabe", "ab", n1rb, n1re, 0, true)
+ *   returns true,
+ *   n1rb is set to "c",
+ *   n1re is set to "e".
+ *
+ * componentContainsBase(y, str.substr(y,0,5), n1rb, n1re, -1, true)
+ *   returns true,
+ *   n1re is set to str.substr(y,5,str.len(y)).
+ *
+ *
+ * Notice that this function may return false when it cannot compute a
+ * remainder when it otherwise would have returned true. For example:
+ *
+ * componentContainsBase(y, str.substr(y,x,z), n1rb, n1re, 0, false)
+ *   returns true.
+ *
+ * Hence, we know that str.substr(y,x,z) is contained in y. However:
+ *
+ * componentContainsBase(y, str.substr(y,x,z), n1rb, n1re, 0, true)
+ *   returns false.
+ *
+ * The reason is since computeRemainder=true, it must be that
+ *   y = str.++( n1rb, str.substr(y,x,z), n1re )
+ * for some n1rb, n1re. However, to construct such n1rb, n1re would require
+ * e.g. the terms:
+ *   y = str.++( ite( x+z < 0 OR x < 0, "", str.substr(y,0,x) ),
+ *               str.substr(y,x,z),
+ *               ite( x+z < 0 OR x < 0, y, str.substr(y,x+z,len(y)) ) )
+ *
+ * Since we do not wish to introduce ITE terms in the rewriter, we instead
+ * return false, indicating that we cannot compute the remainder.
+ */
+static bool componentContainsBase(
+    Node n1, Node n2, Node& n1rb, Node& n1re, int dir, bool computeRemainder);
+/** strip constant endpoints
+ * This function is used when rewriting str.contains( t1, t2 ), where
+ * n1 is the vector form of t1
+ * n2 is the vector form of t2
+ *
+ * It modifies n1 to a new vector n1' such that:
+ * (1) str.contains( str.++( n1 ), str.++( n2 ) ) is equivalent to
+ * str.contains( str.++( n1' ), str.++( n2 ) )
+ * (2) str.++( n1 ) = str.++( nb, n1', ne )
+ *
+ * "dir" is the direction in which we can modify n1:
+ * if dir = 1, then we allow dropping components from the front of n1,
+ * if dir = -1, then we allow dropping components from the back of n1,
+ * if dir = 0, then we allow dropping components from either.
+ *
+ * It returns true if n1 is modified.
+ *
+ * For example:
+ * stripConstantEndpoints({ "ab", x, "de" }, { "c" }, {}, {}, 1)
+ *   returns true,
+ *   n1 is updated to { x, "de" }
+ *   nb is updated to { "ab" }
+ * stripConstantEndpoints({ "ab", x, "de" }, { "bd" }, {}, {}, 0)
+ *   returns true,
+ *   n1 is updated to { "b", x, "d" }
+ *   nb is updated to { "a" }
+ *   ne is updated to { "e" }
+ * stripConstantEndpoints({ "ad", substr("ccc",x,y) }, { "d" }, {}, {}, -1)
+ *   returns true,
+ *   n1 is updated to {"ad"}
+ *   ne is updated to { substr("ccc",x,y) }
+ */
+static bool stripConstantEndpoints(std::vector<Node>& n1,
+                            std::vector<Node>& n2,
+                            std::vector<Node>& nb,
+                            std::vector<Node>& ne,
+                            int dir = 0);
+   
   /**
    * Checks whether a string term `a` is entailed to contain or not contain a
    * string term `b`.
